+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `tec`: Rename project to `tec`
 - `cli`: Use C standard boolean header
 - `tec`: Rename main file from `cli` to `tec`
+- `tec`: Change version output and CLI option
 
 
 ## [0.1.2] - (2025-08-06)